+++ packages/app/src/ui/atoms/link-button/LinkButton.stories.tsx
@@ -2,11 +2,11 @@ import { StoryGrid } from '@sb/components/StoryGrid'
 import type { Meta, StoryObj } from '@storybook/react'
 import { ChevronRightIcon, PlusIcon } from 'lucide-react'
 import { withRouter } from 'storybook-addon-remix-react-router'
-import { ButtonIcon } from '../../button/Button'
+import { ButtonIcon } from '../button/Button'
 import { LinkButton } from './LinkButton'
 
 const meta: Meta<typeof LinkButton> = {
-  title: 'Components/Atoms/New/LinkButton',
+  title: 'Components/Atoms/LinkButton',
   decorators: [withRouter()],
   component: (args) => (
     <StoryGrid className="grid-cols-3">
